requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-restful==0.3.10
Flask-SQLAlchemy==3.1.1
SQLAlchemy>=2.0
PyMySQL==1.1.1
marshmallow==3.21.1
mysql-connector-python==8.4.0
python-dotenv==1.0.1
